fixed bugs

+++ frontend/src/page/history/History.js
@@ -1,5 +1,6 @@
 // Libraries
 import React, { useState, useEffect, Fragment } from "react";
+import swal from "sweetalert";
 // API
 import {
   deleteUserIncome,
@@ -47,17 +48,34 @@ function History() {
 
     // DELETE items
     function deleteItem(userID, subID, type) {
+      swal({
+        title: "Ar tikrai norite ištrinti?",
+        icon: "warning",
+        buttons: ["Atšaukti", "Gerai"],
+      }).then((isConfirm) => {
+        if (isConfirm) {
+          if (type === "income") {
+            deleteUserIncome(users._id, subID, userID).then(() =>
+              updateUserData(users._id)
+            );
+          } else if (type === "expenses") {
+            deleteUserExpenses(users._id, subID, userID).then(() =>
+              updateUserData(users._id)
+            );
+          }
+        }
+      });
       // Incomes
-      if (type === "income") {
-        deleteUserIncome(users._id, subID, userID).then(() => {
-          updateUserData(users._id);
-        });
-        // Expenses
-      } else {
-        deleteUserExpenses(users._id, subID, userID).then(() => {
-          updateUserData(users._id);
-        });
-      }
+      // if (type === "income") {
+      //   deleteUserIncome(users._id, subID, userID).then(() => {
+      //     updateUserData(users._id);
+      //   });
+      //   // Expenses
+      // } else {
+      //   deleteUserExpenses(users._id, subID, userID).then(() => {
+      //     updateUserData(users._id);
+      //   });
+      // }
     }
 
     // Edit button
